fix: update modbus operation

+++ port/modbus.cpp
@@ -42,7 +42,7 @@ void ModbusPort::closePort() {
     m_serial.close();
 }
 
-bool ModbusPort::readModbusData(int startAdd, quint16 numbers, float & val) {
+bool ModbusPort::readModbusData(int startAdd, int numbers, float & val) {
     qDebug("modbus send read message!");
     //读取modbus设备各寄存器数据
     //typeNum:1_线圈 2_离散输入 3_保持 4_输入
@@ -80,7 +80,7 @@ bool ModbusPort::readModbusData(int startAdd, quint16 numbers, float & val) {
 
 //对modbus设备各寄存器写入数据
 //typeNum:1_线圈 2_保持 (这两类寄存器可读可写,其余的只读)
-bool ModbusPort::writeModbusData(int startAdd, uint32_t writeNum, float val)
+bool ModbusPort::writeModbusData(int startAdd, int writeNum, float val)
 {
     return true;
 }
