+++ port/modbus.h
@@ -33,8 +33,8 @@ public:
     virtual void closePort();
     // virtual int sendMsg(int idx);
 
-    virtual bool readModbusData(int startAdd, quint16 numbers, float &val);
-    virtual bool writeModbusData(int startAdd, uint32_t writeNum, float val);
+    virtual bool readModbusData(int startAdd, int numbers, float &val);
+    virtual bool writeModbusData(int startAdd, int writeNum, float val);
     virtual bool waitDataReady();
  
  signals:
