+++ port/modbus_tcp.h
@@ -25,8 +25,8 @@ public:
     virtual void closePort();
     // virtual int sendMsg(int idx);
 
-    virtual bool readModbusData(int startAdd, quint16 numbers, float &val);
-    virtual bool writeModbusData(int startAdd, quint16 numbers, float val);
+    virtual bool readModbusData(int startAdd, int numbers, float &val);
+    virtual bool writeModbusData(int startAdd, int numbers, float val);
     virtual bool waitDataReady();
 
 //  signals:
